Update scaffolding versions
