+++ pkg/plugins/hybrid/v1alpha/scaffolds/internal/templates/main.go
@@ -224,11 +224,13 @@ func init() {
 
 func main() {
 {{- if not .ComponentConfig }}
-	var metricsAddr string
-	var enableLeaderElection bool
-	var probeAddr string
-	var leaderElectionID string
-	var watchesPath string
+	var (
+		metricsAddr          string
+		leaderElectionID     string
+		watchesPath          string
+		probeAddr            string
+		enableLeaderElection bool
+	)
 	
 	flag.StringVar(&metricsAddr, "metrics-bind-address", ":8080", "The address the metric endpoint binds to.")
 	flag.StringVar(&probeAddr, "health-probe-bind-address", ":8081", "The address the probe endpoint binds to.")
